Consolidate datasource testing mechanism
